Agrego put y valido delte de pacientes

+++ backend/paciente.js
@@ -34,22 +34,51 @@ router.post("/",
 
         });
 
-router.put("/", (req, res) => {
+router.put("/",verificarAutenticacion, validarId,
+    body("nombre", "Nombre no valido").isLength({max:20}),
+    body("apellido", "Apellido no valido").isLength({max:20}),
+    body("dni", "DNI no valido").isLength({min:7, max:8}),
+    body("fecha_nacimiento", "Fecha de nacimiento no valida").isDate(),
+    body("obra_social", "Obra social no valida").isLength({max:20}),
+    verificarValidaciones, async (req, res) => {
 
-    const validarExistencia = "SELECT * FROM "
+    const id = Number(req.params.id);
+
+    const [validar] = await db.execute("SELECT * FROM paciente WHERE id = ?", [id]);
+
+    if (validar.length === 0){
+        return res.status(404).json({success: false, message: "Paciente no encontrado"});
+    }
+
+    const {nombre, apellido, dni, fecha_nacimiento, obra_social} = req.body;
+
+    const [validarExisteni] = await db.execute("SELECT * FROM paciente WHERE dni = ? AND id != ?", [dni, id]);
+
+    if (validarExisteni.length > 0){
+        return res.status(400).json({success: false, message: "DNI ya registrado"});
+
+    }
+
+    const [result] = await db.execute("UPDATE paciente SET nombre = ?, apellido = ?, dni = ?, fecha_nacimiento = ?, obra_social = ? WHERE id = ?", [nombre, apellido, dni, fecha_nacimiento, obra_social, id]);
+
+    res.json({success: true, data: {id, nombre, apellido, dni, fecha_nacimiento, obra_social}});
 
 });
 
 router.delete("/:id",verificarAutenticacion, validarId, verificarValidaciones, async (req, res) => {
 
-    
-
     const id = Number(req.params.id);
+
+    const [validar] = await db.execute("SELECT * FROM paciente WHERE id = ?", [id]);
+
+    if (validar.length === 0){
+        return res.status(404).json({success: false, message: "Paciente no encontrado"});
+    }
+
 
     await db.execute("DELETE FROM paciente WHERE id = ?", [id]);
 
     res.json({success: true, message: "Paciente eliminado"});
-
 
 });
